refactor: group LoginController dependencies in a single file
