+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:56456/","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:56497/","mongoDBName":"jest"}
