+++ src/presentation/controller/login/login-protocols.ts
@@ -1,0 +1,3 @@
+export * from '@presentation/protocols' // import generic protocols for controllers
+export * from '@presentation/protocols/email-validator.interface'; // this is an exclusive dependency of signup controller
+export * from '@domain/usecases/authentication';
